Move code re CNV data from Ag3 to AnophelesCnvData (except CNV freq)
